fix sortByText
